we can do it

+++ UI/Head.h
@@ -10,4 +10,4 @@ void body();
 void footer(string role, int button_num);
 void write_frame(int num, string statu, string& lastinput, string& functions);
 void buttons(string& functions);
-void id_write(int num, int& id, string& functions);
+void id_write(int num, double& id, string& functions);
